standardized sign-in form error messages
