+++ src/components/shared/ErrorMsg.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export default styled.p`
+  color: #f44336;
+  font-size: 12px;
+  margin: 10px 0;
+  text-align: left;
+  width: 100%;
+`;
