+++ src/pages/SignIn/index.js
@@ -1,93 +1,8 @@
-import { useState, useContext } from "react";
-import { toast } from "react-toastify";
 
-import AuthLayout from "../../layouts/Auth";
-
-import Input from "../../components/Form/Input";
-import Button from "../../components/Form/Button";
-import Link from "../../components/Link";
-import { Row, Title, Label } from "../../components/Auth";
-
-import EventInfoContext from "../../contexts/EventInfoContext";
-import UserContext from "../../contexts/UserContext";
-
-import useApi from "../../hooks/useApi";
+import SignInForm from "../../components/SignInForm/SignInForm";
 
 export default function SignIn() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [loadingSignIn, setLoadingSignIn] = useState(false);
-
-  const api = useApi();
-
-  const { eventInfo } = useContext(EventInfoContext);
-  const { setUserData } = useContext(UserContext);
-
-  function submit(event) {
-    event.preventDefault();
-    setLoadingSignIn(true);
-
-    api.auth
-      .signIn(email, password)
-      .then((response) => {
-        setLoadingSignIn(false);
-        setUserData(response.data);
-      })
-      .catch((error) => {
-        /* eslint-disable-next-line no-console */
-        console.error(error.response);
-        setLoadingSignIn(false);
-
-        const { status } = error.response;
-
-        if (status === 401) {
-          toast("E-mail ou senha incorreta");
-          return;
-        }
-  
-        toast("Não foi possível conectar ao servidor!");
-      })
-      .then(() => {
-        setLoadingSignIn(false);
-      });
-  }
-
   return (
-    <AuthLayout background={eventInfo.backgroundImage}>
-      <Row>
-        <img src={eventInfo.logoImage} alt="Event Logo" />
-        <Title>{eventInfo.eventTitle}</Title>
-      </Row>
-      <Row>
-        <Label>Entrar</Label>
-        <form onSubmit={submit}>
-          <Input
-            label="E-mail"
-            type="text"
-            fullWidth
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-          />
-          <Input
-            label="Senha"
-            type="password"
-            fullWidth
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-          />
-          <Button
-            type="submit"
-            color="primary"
-            fullWidth
-            disabled={loadingSignIn}
-          >
-            {loadingSignIn ? "Carregando..." : "Entrar"}
-          </Button>
-        </form>
-      </Row>
-      <Row>
-        <Link to="/enroll">Não possui login? Inscreva-se</Link>
-      </Row>
-    </AuthLayout>
+    <SignInForm />
   );
 }
